Fit LoRA training on Kaggle's T4s: fp16, split encoders to GPU 1
A live run OOM'd at text_encoder.to(device) before step 1: the frozen
transformer loaded in fp32 (~6.8GB) plus T5-XXL (~9.5GB) exceeded a single
14.5GB T4 on weights alone.

- Load frozen transformer weights in the training dtype; LoRA params are
  still upcast to fp32 by cast_training_params.
- Train in fp16 instead of bf16. T4s lack native bf16, and under bf16
  PyTorch's memory-efficient SDPA kernel is unavailable on them, so
  attention falls back to materializing the full matrix -- the same ~70GB
  allocation seen in the inference studio at 480x720x49 frames.
- Training is single-process, so put T5 (bf16, fp16-overflow-prone) and
  the VAE (fp32, same reason) on the second GPU and leave GPU 0 to the
  transformer; with one GPU, run T5 on CPU instead.
- Disable in-training validation: it wrapped the live training transformer
  in a new pipeline with enable_model_cpu_offload() (leaving it on CPU
  afterwards) and reloaded a second T5 copy. Guarded and try/excepted so it
  can't kill a multi-hour run if re-enabled.

+++ training/kaggle_entrypoint.py
@@ -135,7 +135,7 @@ def main():
     run([
         "accelerate", "launch",
         "--num_processes", "1",
-        "--mixed_precision", "bf16",
+        "--mixed_precision", "fp16",
         "training/train_lora.py",
         "--task", task,
         "--config", "configs/training_config.yaml",
